Bit of a refactor
