+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/Section.tsx
@@ -4,8 +4,7 @@ import { useTranslation } from "@i18n/client";
 import { FormElementTypes } from "@lib/types";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { useTreeRef } from "@formBuilder/components/shared/right-panel/treeview/provider/TreeRefProvider";
-import { useState } from "react";
-import { useGroupStore } from "@formBuilder/components/shared/right-panel/treeview/store/useGroupStore";
+import { SectionTitle } from "./SectionTitle";
 
 export const Section = ({ groupId }: { groupId: string }) => {
   const { t } = useTranslation("form-builder");
@@ -16,31 +15,12 @@ export const Section = ({ groupId }: { groupId: string }) => {
 
   const { wrapper } = useTreeRef();
   const { handleAddElement } = useHandleAdd(wrapper?.current);
-  const updateGroupName = useGroupStore((state) => state.updateGroupName);
-  const [editing, setEditing] = useState(false);
 
-  const groupName = groups?.[groupId]?.name;
+  const groupName = groups?.[groupId]?.name || "";
 
   if (groupId === "start" || groupId === "end") {
     return null;
   }
-
-  const handleOnBlur = (e: React.FocusEvent) => {
-    saveGroupName(e.currentTarget.textContent || "");
-  };
-
-  const handleOnKeyUp = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      saveGroupName(e.currentTarget.textContent || "");
-    }
-  };
-
-  const saveGroupName = (groupName: string) => {
-    updateGroupName({ id: groupId, name: groupName });
-    setEditing(false);
-    wrapper?.current?.updateItem(groupId, groupName);
-  };
 
   return (
     <>
@@ -53,14 +33,7 @@ export const Section = ({ groupId }: { groupId: string }) => {
         />
       </div>
       <div className="flex max-w-[800px] justify-center rounded-t-lg border-1 border-slate-700 p-2">
-        <h4
-          onDoubleClick={() => setEditing(true)}
-          contentEditable={editing}
-          onBlur={handleOnBlur}
-          onKeyUp={handleOnKeyUp}
-        >
-          {groupName}
-        </h4>
+        <SectionTitle groupName={groupName} groupId={groupId} />
       </div>
     </>
   );
